clean up node.js instrumentation
